Regroup server actions
