+++ server/availability-slot/schema.tsx
@@ -1,0 +1,6 @@
+export type SlotUpsert = {
+  id: string;
+  start: Date;
+  end: Date;
+  label?: string | null;
+};
