+++ server/payment-settings/queries.tsx
@@ -1,5 +1,4 @@
 "use server";
-
 import { db } from "@/lib/db";
 import { paymentSettings } from "@/db/schema";
 import type { PaymentSettings } from "@/db/schema";
