+++ server/payment-settings/schema.tsx
@@ -1,0 +1,6 @@
+export type PaymentSettingsInput = {
+  iban: string;
+  bic?: string;
+  beneficiaryName: string;
+  paymentNote?: string;
+};
