+++ server/user-note/schema.tsx
@@ -1,0 +1,6 @@
+export type UserNotePayload = {
+  id?: string;
+  userId: string;
+  date: Date;
+  note: string;
+};
